Refactor cart functionality and navigation
This commit introduces several changes to enhance cart management and navigation:

- Updated `CartDao` to use `OnConflictStrategy.REPLACE` for adding items, preventing duplicates.
- Modified `getCartItemByProductId` in `CartDao`, `ProductRepository`, and `ProductRepositoryImpl` to return a nullable `CartItem?`.
- In `CartScreenSuccessContent`, the total calculation now correctly casts to `Float`.
- Added `CartScreen` to `NavGraph` and `Destinations`.
- `CartViewModel` now initializes cart items by calling `getCartItems()` in its `init` block.
- Implemented `onAddToCart` in `CategoryProductViewModel` to add or update cart items and navigate to `CartScreen`.

+++ app/src/main/java/com/example/fakestoreappproject/ui/screens/CartScreen.kt
@@ -81,7 +81,7 @@ private fun CartScreenContent(state: CartState, onRemove: (CartItem) -> Unit) {
 
 @Composable
 fun CartScreenSuccessContent(cartItems: List<CartItem>, onRemove: (CartItem) -> Unit) {
-    val total = cartItems.sumOf { it.product.price * it.quantity }
+    val total = cartItems.sumOf { it.product.price * it.quantity }.toFloat()
 
     Scaffold(
         containerColor = Color(0xFFE8EAF6),
